# docs/api/controller.openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "BlindPass Local Controller API",
    "version": "0.1.0",
    "description": "P03 fleet controller contract, including reviewed node, workload, policy, approval, grant, operation and node-channel schemas. Paths marked x-blindpass-status=planned are not mounted yet. This file uses JSON-compatible YAML 1.2 syntax so the checked-in type generator can use Node's standard library without installing a YAML parser.",
    "x-blindpass-ct19": "adopted",
    "x-blindpass-legacy-route-count": 12
  },
  "servers": [
    { "url": "http://127.0.0.1:3200", "description": "Local controller" }
  ],
...
  "paths": {
...
    "/api/v2/agents/token": {
      "post": {
        "operationId": "mintAgentToken",
        "tags": ["Compatibility"],
        "x-blindpass-contract": "legacy-machine",
        "summary": "Exchange an agent bootstrap key for a short-lived agent token",
        "description": "Accepts the bootstrap key as a Bearer credential or in x-agent-api-key. Keys are hashed at rest and invalidated on rotation or revocation. The retained IP rate limit defaults to five requests per minute; hosted paid-tier quota behavior is excluded.",
        "security": [{ "agentBearer": [] }, { "agentApiKey": [] }],
        "responses": {
          "200": { "description": "Agent token minted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentTokenResponse" } } } },
          "401": { "$ref": "#/components/responses/LegacyError" },
          "429": { "$ref": "#/components/responses/LegacyError" }
        }
      }
    },
...
    "/api/v3/node/poll": {
      "post": {
        "operationId": "pollNodeChannel",
        "tags": ["Node channel"],
        "security": [{ "nodeBearer": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NodePollInput" } } } },
        "responses": { "200": { "description": "At most 50 signed documents, and at most about 512 KiB of document text beyond the first (unanswered documents stay unacknowledged for the next poll); an empty bounded long poll carries controller time. Every document is capped at 64 KiB except a provisioning_delivery, which may carry a sealed 64 KiB Source in up to 128 KiB", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NodePollResponse" } } } }, "400": { "$ref": "#/components/responses/AdminError" }, "401": { "$ref": "#/components/responses/AdminError" }, "503": { "$ref": "#/components/responses/AdminError" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "agentBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "nodeBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "Short-lived 15-minute node session bound to an approved node key and audience blindpass-node" },
      "externalWorkloadBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "Configured external issuer with pinned issuer, audience and JWKS validation" },
      "agentApiKey": { "type": "apiKey", "in": "header", "name": "x-agent-api-key" },
      "browserSignedLink": { "type": "apiKey", "in": "query", "name": "sig", "description": "Expiry-bound HMAC signature scoped to the route operation" },
      "bootstrapCapability": { "type": "apiKey", "in": "header", "name": "x-blindpass-bootstrap-token" },
      "adminSession": { "type": "apiKey", "in": "cookie", "name": "bp_session", "description": "Local operator session cookie. While the operator must change a temporary password (bootstrap, local reset or test fixture), every administrative route other than session read, refresh, logout and change-password answers 403 with error password_change_required." },
      "adminRefresh": { "type": "apiKey", "in": "cookie", "name": "bp_refresh" },
      "csrfCookie": { "type": "apiKey", "in": "cookie", "name": "bp_csrf" },
      "desktopBearer": { "type": "http", "scheme": "bearer", "description": "P04-D3 desktop approval-app access token from a kind desktop login or refresh, sent without Origin. Accepted only by session read, refresh and logout and the /api/v3/approvals routes; every other route answers 401. It expires 20 minutes after the login or refresh that issued it, as well as on logout, idle timeout, password change, reset, removal and family revocation. Unsafe methods need no CSRF value because nothing is ambient." },
      "fleetProvisioningCapability": { "type": "apiKey", "in": "query", "name": "sig", "description": "Fleet Source link capability: expiry and a scope-bound (metadata or submit) HMAC in a fleet-specific derived signing domain, distinct from browserSignedLink. It never authorizes by itself: the link's named operator session, role and current authority are also required on every call." },
      "seedToken": { "type": "apiKey", "in": "header", "name": "x-blindpass-seed-token" }
    },
...
    "responses": {
      "LegacyError": { "description": "Legacy route-specific response body preserved from the SPS snapshot", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LegacyError" } } } },
      "AdminError": { "description": "Local administration error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminError" } } } },
...
    },
    "schemas": {
...
      "LegacyError": { "type": "object", "properties": { "code": { "type": "string" }, "error": { "type": "string" }, "message": { "type": "string" }, "statusCode": { "type": "integer" } }, "additionalProperties": true },
...
      "AdminError": { "type": "object", "properties": { "error": { "type": "string" }, "message": { "type": "string" } }, "required": ["error"], "additionalProperties": false },
      "Agent": { "type": "object", "properties": { "id": { "type": "string" }, "agent_id": { "type": "string" }, "workspace_id": { "type": "string" }, "display_name": { "type": "string" }, "status": { "type": "string", "enum": ["active", "revoked"] }, "created_at": { "type": "string", "format": "date-time" }, "revoked_at": { "type": ["string", "null"], "format": "date-time" } }, "required": ["id", "agent_id", "workspace_id", "display_name", "status", "created_at", "revoked_at"], "additionalProperties": false },
      "AgentTokenResponse": { "type": "object", "properties": { "access_token": { "type": "string" }, "access_token_expires_at": { "type": "integer" }, "agent": { "$ref": "#/components/schemas/Agent" } }, "required": ["access_token", "access_token_expires_at", "agent"], "additionalProperties": false },
...
      "SignedDocument": { "type": "object", "properties": { "kind": { "type": "string", "enum": ["registration", "policy_snapshot", "grant", "revocation", "node_revocation", "node_key_rotation", "time_reply", "application_ack", "operation_result", "operation_closed", "recipient_offer", "provisioning_delivery", "audit_event"] }, "v": { "type": "integer", "const": 1 }, "body": { "type": "object", "additionalProperties": true }, "sig": { "type": "string", "pattern": "^[A-Za-z0-9_-]{86}$" }, "kid": { "type": "string" }, "epoch": { "type": "integer", "minimum": 1 } }, "required": ["kind", "v", "body", "sig", "kid", "epoch"], "additionalProperties": false },
...
      "NodePollInput": { "type": "object", "properties": { "ack_seq": { "type": ["integer", "null"], "minimum": 0 }, "health": { "type": "object", "additionalProperties": true }, "time_challenge": { "type": "string", "pattern": "^[A-Za-z0-9_-]{43}$", "description": "One-use suspend-aware broker challenge; when provided the response includes a signed time_reply." } }, "additionalProperties": false },
      "NodeInboxDocument": { "type": "object", "properties": { "seq": { "type": "integer", "minimum": 1 }, "envelope": { "$ref": "#/components/schemas/SignedDocument" } }, "required": ["seq", "envelope"], "additionalProperties": false },
      "NodePollResponse": { "type": "object", "properties": { "documents": { "type": "array", "items": { "$ref": "#/components/schemas/NodeInboxDocument" }, "maxItems": 50 }, "highest_seq": { "type": ["integer", "null"], "minimum": 1 }, "server_time_ms": { "type": "integer" }, "time_reply": { "oneOf": [{ "$ref": "#/components/schemas/SignedDocument" }, { "type": "null" }], "description": "Signed reply bound to time_challenge, node ID and current issuer epoch; null when no challenge was supplied. The body carries challenge_received_at_ms (controller time when the poll arrived) and controller_time_ms (response time); the broker counts any round trip beyond that controller-measured hold as elapsed controller time." } }, "required": ["documents", "highest_seq", "server_time_ms", "time_reply"], "additionalProperties": false },
...
    }
  }
}